utils.py
import streamlit as st
import base64
from google.cloud import secretmanager
import os

def display_instructions():
    # Markdown with some basic CSS styles for the box
    box_css = """
    <style>
        .instructions-box {
            background-color: #f0f0f0;
            border: 1px solid #ddd;
            border-radius: 5px;
            padding: 20px;
        }
    </style>
    """

    st.sidebar.markdown(box_css, unsafe_allow_html=True)

    st.sidebar.markdown(
        """
    <div class="instructions-box">
        
    ### Instructions
    You can exploit this ReAct-based assistant via prompt 
    injection to get two flags:

    - You'll obtain the first flag by accessing the transactions for user with ID 2
    - The second flag is DocBrown's password

    To help you finish the challenge, we suggest you familiarize yourself with the techniques 
    described <a href="https://labs.withsecure.com/publications/llm-agent-prompt-injection" target="_blank">here</a> 
    and <a href="https://youtu.be/43qfHaKh0Xk" target="_blank">here</a>.

    </div>

    You'll also find the database schema to be useful:

    """,
        unsafe_allow_html=True,
    )

    if st.sidebar.button('Show database schema', use_container_width=True):
        st.sidebar.info('Users(userId,username,password)\n\nTransactions(transactionId,username,reference,recipient,amount)')



# Function to convert image to base64
def get_image_base64(path):
    with open(path, "rb") as image_file:
        encoded_string = base64.b64encode(image_file.read()).decode()
    return encoded_string

def get_secret(secret_id: str, version_id: str = "latest") -> str:
    """
    Retrieve a secret from Google Cloud Secret Manager.
    
    Args:
        secret_id: The ID of the secret to retrieve
        version_id: The version of the secret to retrieve (default: "latest")
    
    Returns:
        The secret value as a string
    """
    try:
        project_id = os.getenv("GOOGLE_CLOUD_PROJECT")
        if not project_id:
            raise ValueError("GOOGLE_CLOUD_PROJECT environment variable not set")

        client = secretmanager.SecretManagerServiceClient()
        name = f"projects/{project_id}/secrets/{secret_id}/versions/{version_id}"
        response = client.access_secret_version(request={"name": name})
        return response.payload.data.decode("UTF-8")
    except Exception as e:
        st.error(f"Error retrieving secret: {str(e)}")
        raise

def display_logo():
    # Convert your image
    image_base64 = get_image_base64("labs-logo.png")

    # URL of the company website
    url = 'https://labs.withsecure.com/'

    # HTML for centered image with hyperlink
    html_string = f"""
    <div style="display:flex; justify-content:center;">
        <a href="{url}" target="_blank">
        <img src="data:image/png;base64,{image_base64}" width="150px">
        </a>
    </div>
    """
    # Display the HTML in the sidebar
    st.sidebar.markdown(html_string, unsafe_allow_html=True)
