requirements.txt
streamlit
langchain==0.0.306
openai==0.28.1
google-cloud-secret-manager>=2.16.1
python-dotenv
